% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chr_vec_int}
\alias{chr_vec_int}
\title{turn a string (of elements separated by space) into vector of integers}
\usage{
chr_vec_int(x)
}
\description{
turn a string (of elements separated by space) into vector of integers
}

