% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battleship.R
\name{create_gameboards}
\alias{create_gameboards}
\title{create the hypothesis space of 1.6 million game boards}
\usage{
create_gameboards(x)
}
\description{
create the hypothesis space of 1.6 million game boards
}

