% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{named_list}
\alias{named_list}
\title{list that automatically names its elements}
\usage{
named_list(...)
}
\description{
list that automatically names its elements
}
\references{
http://stackoverflow.com/a/21059868
}

