% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUESTIONS-package.R
\docType{package}
\name{QUESTIONS-package}
\alias{QUESTIONS}
\alias{QUESTIONS-package}
\title{QUESTIONS: QUESTIONS: Quantifying Expression Spillover Trajectories in Omic Network Series}
\description{
QUESTIONS (Quantifying Expression Spillover Trajectories in Omic Network Series) is an R package for modeling signal propagation in dynamic biological networks, particularly those derived from WGCNA. It extends the spillover framework to support timecourse analyses, pruning of dense network matrices, and interpretation of gene- and module-level influence across conditions. QUESTIONS includes tools for pruning, row normalization, dynamic thresholding, and per-sample spillover scoring to support network-based interpretation of omics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CarlyBobak/QUESTIONS}
  \item Report bugs at \url{https://github.com/CarlyBobak/QUESTIONS/issues}
}

}
\author{
\strong{Maintainer}: Carly Bobak \email{your-email@dartmouth.edu}

}
\keyword{internal}
